+++ .cargo/config
@@ -1,6 +1,3 @@
-[build]
-target = "wasm32-unknown-unknown"
-
 [target.wasm32-unknown-unknown]
 rustflags = [
   "-C", "link-arg=-zstack-size=32768",
