+++ Cargo.toml
@@ -16,12 +16,19 @@ hex = "0.4.3"
 wee_alloc = "0.4.5"
 
 [dev-dependencies]
-tokio = { version = "1.12.0", features = ["full"] } 
+tokio = { version = "1.12.0", features = ["full"] }
 ethers = "2.0"
 eyre = "0.6.8"
 
 [features]
 export-abi = ["stylus-sdk/export-abi"]
+
+[[bin]]
+name = "stylus-hello-world"
+path = "src/main.rs"
+
+[lib]
+crate-type = ["lib", "cdylib"]
 
 [profile.release]
 codegen-units = 1
